feat: Add research page
This page is accessible at /research as a top-level destination for the
site. The SiteHeader has also been updated. The purpose of the Research
page is to describe what my research foci are at a high-level and
to provide a glanceable account of the work I have done to date.